requirements.txt
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
distlib==0.3.2
filelock==3.0.12
gTTS==2.2.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Pillow==8.4.0
platformdirs==2.0.2
pycparser==2.21
requests==2.26.0
roslibpy==1.2.1
six==1.16.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.6.0
zope.interface==5.4.0
